fix(ui): address egress-toast review findings
- disable Allow/Deny once when the held call has expired, sharing
  isHeldCallStillLive with the Inbox list
- state-driven stack expansion (hover + focus); collapsed rear cards
  are inert and click-proof; aria-live announcements
- prune timers and hidden ids as rows resolve or become own
- drop the stack below the header on mobile; split the toast card
  into its own component
